# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalReports: number;
  resolvedReports: number;
  pendingReports: number;
  inProgressReports?: number;
  activeEmergencies: number;
  resolutionRate: number;
  avgResolutionHours: number;
  topCategory?: string;
  totalCitizens?: number;
}