# lib/api-zod/src/generated/types/sOSInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SOSInput {
  latitude: number;
  longitude: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  contactPhone?: string | null;
}